Abstracted the old private subclasses from the BattleSnakeGame class. Now we have: BattleSnakeGame: Owns the board and listens for requests for the player to make a move. Board: Holds all the details about the game state, such as opponent snake and food locations. Holds a reference to the player snake, which is declared statically elsewhere. Snake: Base snake class, used to represent opponent snakes. Has the required functionality to deserialise opponent snake info. PlayerSnake: Derived from Snake, implements the player snake. Includes functionality to deserialise the player snake details during a game and the player snake personalisation details like color and head/tail type.
Updated the BattleSnakeController to include a POST route for setting the player snake personalisation from a web page.

Added wwwroot, index.html and bs.js. These implement the controls for a user to set their snake personalisation. Upon setting preferences these are sent to a route in the Battlesnake api where the parameters are applied to the static player snake.

Next to do is further cleasnup of the "framework" code and then implement actual logic for the player snake.