Cleaned up program structure.
Program.cs owns a list of battlesnake games.
BattleSnakeGame owns the player snake. It also has a static PrototypeSnake (instance of PlayerSnake) which is copied to create each instance of a PlayerSnake per game. The PrototypeSnake is where the global personalisation parameters for the players snake are stored at runtime when set by the player in the browser. At some point these parameters should probably be saved to disk so they are retained between game sessions.

PlayerSnake.cs now has a copy constructor for the purpose of duplicating the ProtoypeSnake.

Cleaned up the battlesnake controller to work with the new program structure.