Moved personalisation read/write into player snake class, although it should only really ever be called on the prototype snake. Probably should redesign the program a little.
Moved ProtoypeSnake out of BattleSnakeGame and into Program. It needs to exist earlier in the program lifecycle as it needs to load it's personalisation before the webserver starts up.

Fixed the read/write of the personalisation details. bs.js was using the wrong key names for some of the json members.